@@ -3,7 +3,7 @@ description: Internal package to create platform-appropriate WS clients..
 version: 1.0.0
 
 environment:
-  sdk: ^3.2.0
+  sdk: ^3.3.0
 
 dependencies:
   # stick to specific version, because following version move to incompatible with current implementation pkg:web
