@@ -3,17 +3,17 @@ description: Internal data package of WebTrit Phone application.
 version: 1.0.0
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   clock: ^1.1.1
-  drift: ^2.14.1
+  drift: ^2.15.0
   io: ^1.0.4
   path: ^1.8.3
-  sqlite3_flutter_libs: ^0.5.18
+  sqlite3_flutter_libs: ^0.5.20
 
 dev_dependencies:
-  build_runner: ^2.4.7
-  drift_dev: ^2.14.1
+  build_runner: ^2.4.8
+  drift_dev: ^2.15.0
   lints: ^3.0.0
-  test: ^1.25.0
+  test: ^1.25.2
