@@ -3,8 +3,8 @@ description: Library to extract/get the application info from Apple App Store an
 version: 1.0.0
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
-  flutter: ^3.16.0
+  sdk: '>=3.3.0 <4.0.0'
+  flutter: ^3.19.0
 
 dependencies:
   flutter:
@@ -15,4 +15,4 @@ dependencies:
 
 dev_dependencies:
   lints: ^3.0.0
-  test: ^1.25.0
+  test: ^1.25.2
