@@ -4,7 +4,7 @@ publish_to: none
 version: 1.0.0
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   freezed_annotation: ^2.4.1
@@ -17,8 +17,8 @@ dependencies:
     path: ../_http_client
 
 dev_dependencies:
-  build_runner: ^2.4.7
-  freezed: ^2.4.6
+  build_runner: ^2.4.8
+  freezed: ^2.4.7
   json_serializable: ^6.7.1
   lints: ^3.0.0
-  test: ^1.25.0
+  test: ^1.25.2
