@@ -6,8 +6,8 @@ description: Demonstrates how to use the webtrit_callkeep plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
-  flutter: ^3.16.0
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ^3.19.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
