@@ -4,15 +4,15 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
-  flutter: ^3.16.0
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ^3.19.0
 
 dependencies:
   flutter:
     sdk: flutter
 
   equatable: ^2.0.5
-  uuid: ^4.2.2
+  uuid: ^4.3.3
 
 dev_dependencies:
   flutter_test:
