@@ -3,8 +3,8 @@ description: Flutter package for processing phone numbers.
 version: 0.0.1
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
-  flutter: ^3.16.0
+  sdk: '>=3.3.0 <4.0.0'
+  flutter: ^3.19.0
 
 dependencies:
   flutter:
