@@ -11,8 +11,8 @@ publish_to: 'none'
 version: 1.1.0+1010000
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
-  flutter: ^3.16.0
+  sdk: '>=3.3.0 <4.0.0'
+  flutter: ^3.19.0
 
 dependencies:
   flutter:
@@ -21,20 +21,20 @@ dependencies:
     sdk: flutter
 
   auto_route: ^7.8.4
-  bloc: ^8.1.2
-  bloc_concurrency: ^0.2.3
+  bloc: ^8.1.3
+  bloc_concurrency: ^0.2.4
   characters: ^1.3.0
   clock: ^1.1.1
   connectivity_plus: ^5.0.2
   cross_file: ^0.3.3
-  device_info_plus: ^9.1.1
+  device_info_plus: ^9.1.2
   equatable: ^2.0.5
-  firebase_analytics: ^10.8.0
+  firebase_analytics: ^10.8.6
   firebase_app_installations: ^0.2.4
-  firebase_core: ^2.24.2
-  firebase_crashlytics: ^3.4.9
-  firebase_messaging: ^14.7.10
-  flutter_bloc: ^8.1.3
+  firebase_core: ^2.25.4
+  firebase_crashlytics: ^3.4.15
+  firebase_messaging: ^14.7.16
+  flutter_bloc: ^8.1.4
   flutter_contacts: ^1.1.7
   flutter_secure_storage: ^9.0.0
   flutter_svg: ^2.0.9
@@ -55,17 +55,17 @@ dependencies:
   meta: any
   package_info_plus: ^5.0.1
   path: ^1.8.3
-  path_provider: ^2.1.1
-  permission_handler: ^11.1.0
+  path_provider: ^2.1.2
+  permission_handler: ^11.3.0
   provider: ^6.1.1
   pub_semver: ^2.1.4
-  share_plus: ^7.2.1
+  share_plus: ^7.2.2
   shared_preferences: ^2.2.2
   stream_transform: ^2.1.0
-  url_launcher: ^6.2.2
+  url_launcher: ^6.2.4
   uuid: ^4.2.2
   validators: ^3.0.0
-  webview_flutter: ^4.4.3
+  webview_flutter: ^4.7.0
   webview_flutter_web: ^0.2.2+4
 
   webtrit_api:
@@ -91,13 +91,13 @@ dev_dependencies:
     sdk: flutter
 
   auto_route_generator: ^7.3.2
-  bloc_test: ^9.1.5
-  build_runner: ^2.4.7
+  bloc_test: ^9.1.6
+  build_runner: ^2.4.8
   flutter_gen_runner: ^5.4.0
   flutter_launcher_icons: ^0.13.1
   flutter_lints: ^3.0.1
-  flutter_native_splash: ^2.3.9
-  freezed: ^2.4.6
+  flutter_native_splash: ^2.3.10
+  freezed: ^2.4.7
   json_serializable: ^6.7.1
 
 flutter:
