@@ -12,23 +12,39 @@ part of 'login_cubit.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 /// @nodoc
 mixin _$LoginState {
-  LoginStep get step => throw _privateConstructorUsedError;
-  LoginStatus get status => throw _privateConstructorUsedError;
+  bool get processing => throw _privateConstructorUsedError;
   Object? get error => throw _privateConstructorUsedError;
-  bool get demo => throw _privateConstructorUsedError;
+  LoginMode? get mode => throw _privateConstructorUsedError;
   String? get coreUrl => throw _privateConstructorUsedError;
   String? get tenantId => throw _privateConstructorUsedError;
-  SessionOtpProvisional? get sessionOtpProvisional =>
+  List<LoginType>? get supportedLoginTypes =>
+      throw _privateConstructorUsedError;
+  (
+    SessionOtpProvisional,
+    DateTime
+  )? get otpSigninSessionOtpProvisionalWithDateTime =>
+      throw _privateConstructorUsedError;
+  bool get passwordSigninPasswordInputObscureText =>
+      throw _privateConstructorUsedError;
+  (
+    SessionOtpProvisional,
+    DateTime
+  )? get signupSessionOtpProvisionalWithDateTime =>
       throw _privateConstructorUsedError;
   String? get token => throw _privateConstructorUsedError;
   UrlInput get coreUrlInput => throw _privateConstructorUsedError;
-  EmailInput get emailInput => throw _privateConstructorUsedError;
-  PhoneInput get phoneInput => throw _privateConstructorUsedError;
-  CodeInput get codeInput => throw _privateConstructorUsedError;
+  UserRefInput get otpSigninUserRefInput => throw _privateConstructorUsedError;
+  CodeInput get otpSigninCodeInput => throw _privateConstructorUsedError;
+  UserRefInput get passwordSigninUserRefInput =>
+      throw _privateConstructorUsedError;
+  PasswordInput get passwordSigninPasswordInput =>
+      throw _privateConstructorUsedError;
+  EmailInput get signupEmailInput => throw _privateConstructorUsedError;
+  CodeInput get signupCodeInput => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
   $LoginStateCopyWith<LoginState> get copyWith =>
@@ -42,18 +58,29 @@ abstract class $LoginStateCopyWith<$Res> {
       _$LoginStateCopyWithImpl<$Res, LoginState>;
   @useResult
   $Res call(
-      {LoginStep step,
-      LoginStatus status,
+      {bool processing,
       Object? error,
-      bool demo,
+      LoginMode? mode,
       String? coreUrl,
       String? tenantId,
-      SessionOtpProvisional? sessionOtpProvisional,
+      List<LoginType>? supportedLoginTypes,
+      (
+        SessionOtpProvisional,
+        DateTime
+      )? otpSigninSessionOtpProvisionalWithDateTime,
+      bool passwordSigninPasswordInputObscureText,
+      (
+        SessionOtpProvisional,
+        DateTime
+      )? signupSessionOtpProvisionalWithDateTime,
       String? token,
       UrlInput coreUrlInput,
-      EmailInput emailInput,
-      PhoneInput phoneInput,
-      CodeInput codeInput});
+      UserRefInput otpSigninUserRefInput,
+      CodeInput otpSigninCodeInput,
+      UserRefInput passwordSigninUserRefInput,
+      PasswordInput passwordSigninPasswordInput,
+      EmailInput signupEmailInput,
+      CodeInput signupCodeInput});
 }
 
 /// @nodoc
@@ -69,33 +96,34 @@ class _$LoginStateCopyWithImpl<$Res, $Val extends LoginState>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? step = null,
-    Object? status = null,
+    Object? processing = null,
     Object? error = freezed,
-    Object? demo = null,
+    Object? mode = freezed,
     Object? coreUrl = freezed,
     Object? tenantId = freezed,
-    Object? sessionOtpProvisional = freezed,
+    Object? supportedLoginTypes = freezed,
+    Object? otpSigninSessionOtpProvisionalWithDateTime = freezed,
+    Object? passwordSigninPasswordInputObscureText = null,
+    Object? signupSessionOtpProvisionalWithDateTime = freezed,
     Object? token = freezed,
     Object? coreUrlInput = null,
-    Object? emailInput = null,
-    Object? phoneInput = null,
-    Object? codeInput = null,
+    Object? otpSigninUserRefInput = null,
+    Object? otpSigninCodeInput = null,
+    Object? passwordSigninUserRefInput = null,
+    Object? passwordSigninPasswordInput = null,
+    Object? signupEmailInput = null,
+    Object? signupCodeInput = null,
   }) {
     return _then(_value.copyWith(
-      step: null == step
-          ? _value.step
-          : step // ignore: cast_nullable_to_non_nullable
-              as LoginStep,
-      status: null == status
-          ? _value.status
-          : status // ignore: cast_nullable_to_non_nullable
-              as LoginStatus,
-      error: freezed == error ? _value.error : error,
-      demo: null == demo
-          ? _value.demo
-          : demo // ignore: cast_nullable_to_non_nullable
+      processing: null == processing
+          ? _value.processing
+          : processing // ignore: cast_nullable_to_non_nullable
               as bool,
+      error: freezed == error ? _value.error : error,
+      mode: freezed == mode
+          ? _value.mode
+          : mode // ignore: cast_nullable_to_non_nullable
+              as LoginMode?,
       coreUrl: freezed == coreUrl
           ? _value.coreUrl
           : coreUrl // ignore: cast_nullable_to_non_nullable
@@ -104,10 +132,25 @@ class _$LoginStateCopyWithImpl<$Res, $Val extends LoginState>
           ? _value.tenantId
           : tenantId // ignore: cast_nullable_to_non_nullable
               as String?,
-      sessionOtpProvisional: freezed == sessionOtpProvisional
-          ? _value.sessionOtpProvisional
-          : sessionOtpProvisional // ignore: cast_nullable_to_non_nullable
-              as SessionOtpProvisional?,
+      supportedLoginTypes: freezed == supportedLoginTypes
+          ? _value.supportedLoginTypes
+          : supportedLoginTypes // ignore: cast_nullable_to_non_nullable
+              as List<LoginType>?,
+      otpSigninSessionOtpProvisionalWithDateTime: freezed ==
+              otpSigninSessionOtpProvisionalWithDateTime
+          ? _value.otpSigninSessionOtpProvisionalWithDateTime
+          : otpSigninSessionOtpProvisionalWithDateTime // ignore: cast_nullable_to_non_nullable
+              as (SessionOtpProvisional, DateTime)?,
+      passwordSigninPasswordInputObscureText: null ==
+              passwordSigninPasswordInputObscureText
+          ? _value.passwordSigninPasswordInputObscureText
+          : passwordSigninPasswordInputObscureText // ignore: cast_nullable_to_non_nullable
+              as bool,
+      signupSessionOtpProvisionalWithDateTime: freezed ==
+              signupSessionOtpProvisionalWithDateTime
+          ? _value.signupSessionOtpProvisionalWithDateTime
+          : signupSessionOtpProvisionalWithDateTime // ignore: cast_nullable_to_non_nullable
+              as (SessionOtpProvisional, DateTime)?,
       token: freezed == token
           ? _value.token
           : token // ignore: cast_nullable_to_non_nullable
@@ -116,17 +159,29 @@ class _$LoginStateCopyWithImpl<$Res, $Val extends LoginState>
           ? _value.coreUrlInput
           : coreUrlInput // ignore: cast_nullable_to_non_nullable
               as UrlInput,
-      emailInput: null == emailInput
-          ? _value.emailInput
-          : emailInput // ignore: cast_nullable_to_non_nullable
+      otpSigninUserRefInput: null == otpSigninUserRefInput
+          ? _value.otpSigninUserRefInput
+          : otpSigninUserRefInput // ignore: cast_nullable_to_non_nullable
+              as UserRefInput,
+      otpSigninCodeInput: null == otpSigninCodeInput
+          ? _value.otpSigninCodeInput
+          : otpSigninCodeInput // ignore: cast_nullable_to_non_nullable
+              as CodeInput,
+      passwordSigninUserRefInput: null == passwordSigninUserRefInput
+          ? _value.passwordSigninUserRefInput
+          : passwordSigninUserRefInput // ignore: cast_nullable_to_non_nullable
+              as UserRefInput,
+      passwordSigninPasswordInput: null == passwordSigninPasswordInput
+          ? _value.passwordSigninPasswordInput
+          : passwordSigninPasswordInput // ignore: cast_nullable_to_non_nullable
+              as PasswordInput,
+      signupEmailInput: null == signupEmailInput
+          ? _value.signupEmailInput
+          : signupEmailInput // ignore: cast_nullable_to_non_nullable
               as EmailInput,
-      phoneInput: null == phoneInput
-          ? _value.phoneInput
-          : phoneInput // ignore: cast_nullable_to_non_nullable
-              as PhoneInput,
-      codeInput: null == codeInput
-          ? _value.codeInput
-          : codeInput // ignore: cast_nullable_to_non_nullable
+      signupCodeInput: null == signupCodeInput
+          ? _value.signupCodeInput
+          : signupCodeInput // ignore: cast_nullable_to_non_nullable
               as CodeInput,
     ) as $Val);
   }
@@ -141,18 +196,29 @@ abstract class _$$LoginStateImplCopyWith<$Res>
   @override
   @useResult
   $Res call(
-      {LoginStep step,
-      LoginStatus status,
+      {bool processing,
       Object? error,
-      bool demo,
+      LoginMode? mode,
       String? coreUrl,
       String? tenantId,
-      SessionOtpProvisional? sessionOtpProvisional,
+      List<LoginType>? supportedLoginTypes,
+      (
+        SessionOtpProvisional,
+        DateTime
+      )? otpSigninSessionOtpProvisionalWithDateTime,
+      bool passwordSigninPasswordInputObscureText,
+      (
+        SessionOtpProvisional,
+        DateTime
+      )? signupSessionOtpProvisionalWithDateTime,
       String? token,
       UrlInput coreUrlInput,
-      EmailInput emailInput,
-      PhoneInput phoneInput,
-      CodeInput codeInput});
+      UserRefInput otpSigninUserRefInput,
+      CodeInput otpSigninCodeInput,
+      UserRefInput passwordSigninUserRefInput,
+      PasswordInput passwordSigninPasswordInput,
+      EmailInput signupEmailInput,
+      CodeInput signupCodeInput});
 }
 
 /// @nodoc
@@ -166,33 +232,34 @@ class __$$LoginStateImplCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? step = null,
-    Object? status = null,
+    Object? processing = null,
     Object? error = freezed,
-    Object? demo = null,
+    Object? mode = freezed,
     Object? coreUrl = freezed,
     Object? tenantId = freezed,
-    Object? sessionOtpProvisional = freezed,
+    Object? supportedLoginTypes = freezed,
+    Object? otpSigninSessionOtpProvisionalWithDateTime = freezed,
+    Object? passwordSigninPasswordInputObscureText = null,
+    Object? signupSessionOtpProvisionalWithDateTime = freezed,
     Object? token = freezed,
     Object? coreUrlInput = null,
-    Object? emailInput = null,
-    Object? phoneInput = null,
-    Object? codeInput = null,
+    Object? otpSigninUserRefInput = null,
+    Object? otpSigninCodeInput = null,
+    Object? passwordSigninUserRefInput = null,
+    Object? passwordSigninPasswordInput = null,
+    Object? signupEmailInput = null,
+    Object? signupCodeInput = null,
   }) {
     return _then(_$LoginStateImpl(
-      step: null == step
-          ? _value.step
-          : step // ignore: cast_nullable_to_non_nullable
-              as LoginStep,
-      status: null == status
-          ? _value.status
-          : status // ignore: cast_nullable_to_non_nullable
-              as LoginStatus,
-      error: freezed == error ? _value.error : error,
-      demo: null == demo
-          ? _value.demo
-          : demo // ignore: cast_nullable_to_non_nullable
+      processing: null == processing
+          ? _value.processing
+          : processing // ignore: cast_nullable_to_non_nullable
               as bool,
+      error: freezed == error ? _value.error : error,
+      mode: freezed == mode
+          ? _value.mode
+          : mode // ignore: cast_nullable_to_non_nullable
+              as LoginMode?,
       coreUrl: freezed == coreUrl
           ? _value.coreUrl
           : coreUrl // ignore: cast_nullable_to_non_nullable
@@ -201,10 +268,25 @@ class __$$LoginStateImplCopyWithImpl<$Res>
           ? _value.tenantId
           : tenantId // ignore: cast_nullable_to_non_nullable
               as String?,
-      sessionOtpProvisional: freezed == sessionOtpProvisional
-          ? _value.sessionOtpProvisional
-          : sessionOtpProvisional // ignore: cast_nullable_to_non_nullable
-              as SessionOtpProvisional?,
+      supportedLoginTypes: freezed == supportedLoginTypes
+          ? _value._supportedLoginTypes
+          : supportedLoginTypes // ignore: cast_nullable_to_non_nullable
+              as List<LoginType>?,
+      otpSigninSessionOtpProvisionalWithDateTime: freezed ==
+              otpSigninSessionOtpProvisionalWithDateTime
+          ? _value.otpSigninSessionOtpProvisionalWithDateTime
+          : otpSigninSessionOtpProvisionalWithDateTime // ignore: cast_nullable_to_non_nullable
+              as (SessionOtpProvisional, DateTime)?,
+      passwordSigninPasswordInputObscureText: null ==
+              passwordSigninPasswordInputObscureText
+          ? _value.passwordSigninPasswordInputObscureText
+          : passwordSigninPasswordInputObscureText // ignore: cast_nullable_to_non_nullable
+              as bool,
+      signupSessionOtpProvisionalWithDateTime: freezed ==
+              signupSessionOtpProvisionalWithDateTime
+          ? _value.signupSessionOtpProvisionalWithDateTime
+          : signupSessionOtpProvisionalWithDateTime // ignore: cast_nullable_to_non_nullable
+              as (SessionOtpProvisional, DateTime)?,
       token: freezed == token
           ? _value.token
           : token // ignore: cast_nullable_to_non_nullable
@@ -213,17 +295,29 @@ class __$$LoginStateImplCopyWithImpl<$Res>
           ? _value.coreUrlInput
           : coreUrlInput // ignore: cast_nullable_to_non_nullable
               as UrlInput,
-      emailInput: null == emailInput
-          ? _value.emailInput
-          : emailInput // ignore: cast_nullable_to_non_nullable
+      otpSigninUserRefInput: null == otpSigninUserRefInput
+          ? _value.otpSigninUserRefInput
+          : otpSigninUserRefInput // ignore: cast_nullable_to_non_nullable
+              as UserRefInput,
+      otpSigninCodeInput: null == otpSigninCodeInput
+          ? _value.otpSigninCodeInput
+          : otpSigninCodeInput // ignore: cast_nullable_to_non_nullable
+              as CodeInput,
+      passwordSigninUserRefInput: null == passwordSigninUserRefInput
+          ? _value.passwordSigninUserRefInput
+          : passwordSigninUserRefInput // ignore: cast_nullable_to_non_nullable
+              as UserRefInput,
+      passwordSigninPasswordInput: null == passwordSigninPasswordInput
+          ? _value.passwordSigninPasswordInput
+          : passwordSigninPasswordInput // ignore: cast_nullable_to_non_nullable
+              as PasswordInput,
+      signupEmailInput: null == signupEmailInput
+          ? _value.signupEmailInput
+          : signupEmailInput // ignore: cast_nullable_to_non_nullable
               as EmailInput,
-      phoneInput: null == phoneInput
-          ? _value.phoneInput
-          : phoneInput // ignore: cast_nullable_to_non_nullable
-              as PhoneInput,
-      codeInput: null == codeInput
-          ? _value.codeInput
-          : codeInput // ignore: cast_nullable_to_non_nullable
+      signupCodeInput: null == signupCodeInput
+          ? _value.signupCodeInput
+          : signupCodeInput // ignore: cast_nullable_to_non_nullable
               as CodeInput,
     ));
   }
@@ -233,35 +327,60 @@ class __$$LoginStateImplCopyWithImpl<$Res>
 
 class _$LoginStateImpl implements _LoginState {
   const _$LoginStateImpl(
-      {required this.step,
-      this.status = LoginStatus.input,
+      {this.processing = false,
       this.error,
-      this.demo = false,
+      this.mode,
       this.coreUrl,
       this.tenantId,
-      this.sessionOtpProvisional,
+      final List<LoginType>? supportedLoginTypes,
+      this.otpSigninSessionOtpProvisionalWithDateTime,
+      this.passwordSigninPasswordInputObscureText = true,
+      this.signupSessionOtpProvisionalWithDateTime,
       this.token,
       this.coreUrlInput = const UrlInput.pure(),
-      this.emailInput = const EmailInput.pure(),
-      this.phoneInput = const PhoneInput.pure(),
-      this.codeInput = const CodeInput.pure()});
+      this.otpSigninUserRefInput = const UserRefInput.pure(),
+      this.otpSigninCodeInput = const CodeInput.pure(),
+      this.passwordSigninUserRefInput = const UserRefInput.pure(),
+      this.passwordSigninPasswordInput = const PasswordInput.pure(),
+      this.signupEmailInput = const EmailInput.pure(),
+      this.signupCodeInput = const CodeInput.pure()})
+      : _supportedLoginTypes = supportedLoginTypes;
 
   @override
-  final LoginStep step;
-  @override
   @JsonKey()
-  final LoginStatus status;
+  final bool processing;
   @override
   final Object? error;
   @override
-  @JsonKey()
-  final bool demo;
+  final LoginMode? mode;
   @override
   final String? coreUrl;
   @override
   final String? tenantId;
+  final List<LoginType>? _supportedLoginTypes;
   @override
-  final SessionOtpProvisional? sessionOtpProvisional;
+  List<LoginType>? get supportedLoginTypes {
+    final value = _supportedLoginTypes;
+    if (value == null) return null;
+    if (_supportedLoginTypes is EqualUnmodifiableListView)
+      return _supportedLoginTypes;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(value);
+  }
+
+  @override
+  final (
+    SessionOtpProvisional,
+    DateTime
+  )? otpSigninSessionOtpProvisionalWithDateTime;
+  @override
+  @JsonKey()
+  final bool passwordSigninPasswordInputObscureText;
+  @override
+  final (
+    SessionOtpProvisional,
+    DateTime
+  )? signupSessionOtpProvisionalWithDateTime;
   @override
   final String? token;
   @override
@@ -269,17 +388,26 @@ class _$LoginStateImpl implements _LoginState {
   final UrlInput coreUrlInput;
   @override
   @JsonKey()
-  final EmailInput emailInput;
+  final UserRefInput otpSigninUserRefInput;
   @override
   @JsonKey()
-  final PhoneInput phoneInput;
+  final CodeInput otpSigninCodeInput;
   @override
   @JsonKey()
-  final CodeInput codeInput;
+  final UserRefInput passwordSigninUserRefInput;
+  @override
+  @JsonKey()
+  final PasswordInput passwordSigninPasswordInput;
+  @override
+  @JsonKey()
+  final EmailInput signupEmailInput;
+  @override
+  @JsonKey()
+  final CodeInput signupCodeInput;
 
   @override
   String toString() {
-    return 'LoginState(step: $step, status: $status, error: $error, demo: $demo, coreUrl: $coreUrl, tenantId: $tenantId, sessionOtpProvisional: $sessionOtpProvisional, token: $token, coreUrlInput: $coreUrlInput, emailInput: $emailInput, phoneInput: $phoneInput, codeInput: $codeInput)';
+    return 'LoginState(processing: $processing, error: $error, mode: $mode, coreUrl: $coreUrl, tenantId: $tenantId, supportedLoginTypes: $supportedLoginTypes, otpSigninSessionOtpProvisionalWithDateTime: $otpSigninSessionOtpProvisionalWithDateTime, passwordSigninPasswordInputObscureText: $passwordSigninPasswordInputObscureText, signupSessionOtpProvisionalWithDateTime: $signupSessionOtpProvisionalWithDateTime, token: $token, coreUrlInput: $coreUrlInput, otpSigninUserRefInput: $otpSigninUserRefInput, otpSigninCodeInput: $otpSigninCodeInput, passwordSigninUserRefInput: $passwordSigninUserRefInput, passwordSigninPasswordInput: $passwordSigninPasswordInput, signupEmailInput: $signupEmailInput, signupCodeInput: $signupCodeInput)';
   }
 
   @override
@@ -287,41 +415,66 @@ class _$LoginStateImpl implements _LoginState {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$LoginStateImpl &&
-            (identical(other.step, step) || other.step == step) &&
-            (identical(other.status, status) || other.status == status) &&
+            (identical(other.processing, processing) ||
+                other.processing == processing) &&
             const DeepCollectionEquality().equals(other.error, error) &&
-            (identical(other.demo, demo) || other.demo == demo) &&
+            (identical(other.mode, mode) || other.mode == mode) &&
             (identical(other.coreUrl, coreUrl) || other.coreUrl == coreUrl) &&
             (identical(other.tenantId, tenantId) ||
                 other.tenantId == tenantId) &&
-            (identical(other.sessionOtpProvisional, sessionOtpProvisional) ||
-                other.sessionOtpProvisional == sessionOtpProvisional) &&
+            const DeepCollectionEquality()
+                .equals(other._supportedLoginTypes, _supportedLoginTypes) &&
+            (identical(other.otpSigninSessionOtpProvisionalWithDateTime,
+                    otpSigninSessionOtpProvisionalWithDateTime) ||
+                other.otpSigninSessionOtpProvisionalWithDateTime ==
+                    otpSigninSessionOtpProvisionalWithDateTime) &&
+            (identical(other.passwordSigninPasswordInputObscureText,
+                    passwordSigninPasswordInputObscureText) ||
+                other.passwordSigninPasswordInputObscureText ==
+                    passwordSigninPasswordInputObscureText) &&
+            (identical(other.signupSessionOtpProvisionalWithDateTime,
+                    signupSessionOtpProvisionalWithDateTime) ||
+                other.signupSessionOtpProvisionalWithDateTime ==
+                    signupSessionOtpProvisionalWithDateTime) &&
             (identical(other.token, token) || other.token == token) &&
             (identical(other.coreUrlInput, coreUrlInput) ||
                 other.coreUrlInput == coreUrlInput) &&
-            (identical(other.emailInput, emailInput) ||
-                other.emailInput == emailInput) &&
-            (identical(other.phoneInput, phoneInput) ||
-                other.phoneInput == phoneInput) &&
-            (identical(other.codeInput, codeInput) ||
-                other.codeInput == codeInput));
+            (identical(other.otpSigninUserRefInput, otpSigninUserRefInput) ||
+                other.otpSigninUserRefInput == otpSigninUserRefInput) &&
+            (identical(other.otpSigninCodeInput, otpSigninCodeInput) ||
+                other.otpSigninCodeInput == otpSigninCodeInput) &&
+            (identical(other.passwordSigninUserRefInput, passwordSigninUserRefInput) ||
+                other.passwordSigninUserRefInput ==
+                    passwordSigninUserRefInput) &&
+            (identical(other.passwordSigninPasswordInput, passwordSigninPasswordInput) ||
+                other.passwordSigninPasswordInput ==
+                    passwordSigninPasswordInput) &&
+            (identical(other.signupEmailInput, signupEmailInput) ||
+                other.signupEmailInput == signupEmailInput) &&
+            (identical(other.signupCodeInput, signupCodeInput) ||
+                other.signupCodeInput == signupCodeInput));
   }
 
   @override
   int get hashCode => Object.hash(
       runtimeType,
-      step,
-      status,
+      processing,
       const DeepCollectionEquality().hash(error),
-      demo,
+      mode,
       coreUrl,
       tenantId,
-      sessionOtpProvisional,
+      const DeepCollectionEquality().hash(_supportedLoginTypes),
+      otpSigninSessionOtpProvisionalWithDateTime,
+      passwordSigninPasswordInputObscureText,
+      signupSessionOtpProvisionalWithDateTime,
       token,
       coreUrlInput,
-      emailInput,
-      phoneInput,
-      codeInput);
+      otpSigninUserRefInput,
+      otpSigninCodeInput,
+      passwordSigninUserRefInput,
+      passwordSigninPasswordInput,
+      signupEmailInput,
+      signupCodeInput);
 
   @JsonKey(ignore: true)
   @override
@@ -332,43 +485,66 @@ class _$LoginStateImpl implements _LoginState {
 
 abstract class _LoginState implements LoginState {
   const factory _LoginState(
-      {required final LoginStep step,
-      final LoginStatus status,
+      {final bool processing,
       final Object? error,
-      final bool demo,
+      final LoginMode? mode,
       final String? coreUrl,
       final String? tenantId,
-      final SessionOtpProvisional? sessionOtpProvisional,
+      final List<LoginType>? supportedLoginTypes,
+      final (
+        SessionOtpProvisional,
+        DateTime
+      )? otpSigninSessionOtpProvisionalWithDateTime,
+      final bool passwordSigninPasswordInputObscureText,
+      final (
+        SessionOtpProvisional,
+        DateTime
+      )? signupSessionOtpProvisionalWithDateTime,
       final String? token,
       final UrlInput coreUrlInput,
-      final EmailInput emailInput,
-      final PhoneInput phoneInput,
-      final CodeInput codeInput}) = _$LoginStateImpl;
+      final UserRefInput otpSigninUserRefInput,
+      final CodeInput otpSigninCodeInput,
+      final UserRefInput passwordSigninUserRefInput,
+      final PasswordInput passwordSigninPasswordInput,
+      final EmailInput signupEmailInput,
+      final CodeInput signupCodeInput}) = _$LoginStateImpl;
 
   @override
-  LoginStep get step;
-  @override
-  LoginStatus get status;
+  bool get processing;
   @override
   Object? get error;
   @override
-  bool get demo;
+  LoginMode? get mode;
   @override
   String? get coreUrl;
   @override
   String? get tenantId;
   @override
-  SessionOtpProvisional? get sessionOtpProvisional;
+  List<LoginType>? get supportedLoginTypes;
+  @override
+  (SessionOtpProvisional, DateTime)?
+      get otpSigninSessionOtpProvisionalWithDateTime;
+  @override
+  bool get passwordSigninPasswordInputObscureText;
+  @override
+  (SessionOtpProvisional, DateTime)?
+      get signupSessionOtpProvisionalWithDateTime;
   @override
   String? get token;
   @override
   UrlInput get coreUrlInput;
   @override
-  EmailInput get emailInput;
+  UserRefInput get otpSigninUserRefInput;
   @override
-  PhoneInput get phoneInput;
+  CodeInput get otpSigninCodeInput;
   @override
-  CodeInput get codeInput;
+  UserRefInput get passwordSigninUserRefInput;
+  @override
+  PasswordInput get passwordSigninPasswordInput;
+  @override
+  EmailInput get signupEmailInput;
+  @override
+  CodeInput get signupCodeInput;
   @override
   @JsonKey(ignore: true)
   _$$LoginStateImplCopyWith<_$LoginStateImpl> get copyWith =>
