@@ -6,21 +6,25 @@ class ContactPhoneTile extends StatelessWidget {
     required this.number,
     required this.label,
     required this.favorite,
+    required this.transfer,
     this.onTap,
     this.onLongPress,
     this.onFavoriteChanged,
     this.onAudioPressed,
     this.onVideoPressed,
+    this.onTransferPressed,
   });
 
   final String number;
   final String label;
   final bool favorite;
+  final bool transfer;
   final GestureTapCallback? onTap;
   final GestureLongPressCallback? onLongPress;
   final ValueChanged<bool>? onFavoriteChanged;
   final VoidCallback? onAudioPressed;
   final VoidCallback? onVideoPressed;
+  final VoidCallback? onTransferPressed;
 
   @override
   Widget build(BuildContext context) {
@@ -34,16 +38,24 @@ class ContactPhoneTile extends StatelessWidget {
             icon: favorite ? const Icon(Icons.star) : const Icon(Icons.star_border),
             onPressed: onFavoriteChanged == null ? null : () => onFavoriteChanged!(!favorite),
           ),
-          IconButton(
-            splashRadius: 24,
-            icon: const Icon(Icons.call),
-            onPressed: onAudioPressed,
-          ),
-          IconButton(
-            splashRadius: 24,
-            icon: const Icon(Icons.videocam),
-            onPressed: onVideoPressed,
-          ),
+          if (transfer)
+            IconButton(
+              splashRadius: 24,
+              icon: const Icon(Icons.phone_forwarded),
+              onPressed: onTransferPressed,
+            )
+          else ...[
+            IconButton(
+              splashRadius: 24,
+              icon: const Icon(Icons.call),
+              onPressed: onAudioPressed,
+            ),
+            IconButton(
+              splashRadius: 24,
+              icon: const Icon(Icons.videocam),
+              onPressed: onVideoPressed,
+            ),
+          ],
         ],
       ),
       title: Text(number),
