@@ -9,15 +9,19 @@ class Actionpad extends StatelessWidget {
   const Actionpad({
     super.key,
     this.video = false,
+    this.transfer = false,
     this.onCallPressed,
     this.onCallLongPress,
+    this.onTransferPressed,
     this.onBackspacePressed,
     this.onBackspaceLongPress,
   });
 
   final bool video;
+  final bool transfer;
   final VoidCallback? onCallPressed;
   final VoidCallback? onCallLongPress;
+  final VoidCallback? onTransferPressed;
   final VoidCallback? onBackspacePressed;
   final VoidCallback? onBackspaceLongPress;
 
@@ -34,11 +38,15 @@ class Actionpad extends StatelessWidget {
       children: [
         const SizedBox(),
         TextButton(
-          onPressed: onCallPressed,
-          onLongPress: onCallLongPress,
-          style: textButtonStyles?.callStart,
+          onPressed: transfer ? onTransferPressed : onCallPressed,
+          onLongPress: transfer ? null : onCallLongPress,
+          style: transfer ? textButtonStyles?.callTransfer : textButtonStyles?.callStart,
           child: Icon(
-            video ? Icons.videocam : Icons.call,
+            transfer
+                ? Icons.phone_forwarded
+                : video
+                    ? Icons.videocam
+                    : Icons.call,
             size: Theme.of(context).textTheme.displayMedium!.fontSize,
           ),
         ),
