@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:bloc/bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:linkify/linkify.dart';
@@ -23,10 +25,9 @@ WebtritApiClient defaultCreateWebtritApiClient(String coreUrl, String tenantId) 
 }
 
 class LoginCubit extends Cubit<LoginState> {
-  LoginCubit(
-    LoginStep step, {
+  LoginCubit({
     @visibleForTesting this.createWebtritApiClient = defaultCreateWebtritApiClient,
-  }) : super(LoginState(step: step));
+  }) : super(const LoginState());
 
   final WebtritApiClientFactory createWebtritApiClient;
 
@@ -37,29 +38,6 @@ class LoginCubit extends Cubit<LoginState> {
   String get defaultTenantId => '';
 
   bool get isDemoModeEnabled => coreUrlFromEnvironment == null;
-
-  void next() {
-    var nextStepIndex = state.step.index + 1;
-    if (nextStepIndex == 1 && state.coreUrl != null) {
-      nextStepIndex++;
-    }
-    emit(state.copyWith(
-      step: LoginStep.values[nextStepIndex],
-      status: LoginStatus.input,
-    ));
-  }
-
-  void back() {
-    var prevStepIndex = state.step.index - 1;
-    if (prevStepIndex == 1 && state.coreUrlInput.value.isEmpty) {
-      prevStepIndex--;
-    }
-    final effectivePrevStep = prevStepIndex >= 0 ? LoginStep.values[prevStepIndex] : state.step;
-    emit(state.copyWith(
-      step: effectivePrevStep,
-      status: LoginStatus.input,
-    ));
-  }
 
   void dismissError() {
     emit(state.copyWith(
@@ -74,243 +52,369 @@ class LoginCubit extends Cubit<LoginState> {
     }
   }
 
-  // LoginModeSelectTab
-
-  void loginModeSelectSubmitter(bool demo) async {
+  Future<void> _verifyCoreVersionAndRetrieveSupportedLoginTypesSubmitted(String coreUrl, String tenantId,
+      [bool demo = false]) async {
     emit(state.copyWith(
-      demo: demo,
+      processing: true,
     ));
 
-    final coreUrl = demo ? demoCoreUrlFromEnvironment : coreUrlFromEnvironment;
-    final tenantId = defaultTenantId;
-
-    if (coreUrl != null) {
-      emit(state.copyWith(
-        status: LoginStatus.processing,
-      ));
-      try {
-        final client = createWebtritApiClient(coreUrl, tenantId);
-        await _verifyCoreVersion(client);
+    try {
+      final client = createWebtritApiClient(coreUrl, tenantId);
+      final supportedLoginTypes = await _verifyCoreVersionAndRetrieveSupportedLoginTypes(client);
+      if (demo) {
+        supportedLoginTypes?.removeWhere((loginType) => loginType != LoginType.signup);
+      }
+      if (supportedLoginTypes != null && supportedLoginTypes.isNotEmpty) {
         emit(state.copyWith(
-          status: LoginStatus.ok,
+          processing: false,
           coreUrl: coreUrl,
           tenantId: tenantId,
+          supportedLoginTypes: supportedLoginTypes,
         ));
-      } catch (e) {
+      } else {
         emit(state.copyWith(
-          status: LoginStatus.input,
-          error: e,
+          processing: false,
+          error: SupportedLoginTypeMissedException(),
         ));
       }
-    } else {
+    } catch (e) {
       emit(state.copyWith(
-        status: LoginStatus.ok,
+        processing: false,
+        error: e,
       ));
     }
   }
 
-  // LoginCoreUrlAssignTab
+  // LoginModeSelect
 
-  void loginCoreUrlAssignCoreUrlInputChanged(String value) {
+  void loginModeSelectSubmitted(LoginMode mode) async {
+    emit(state.copyWith(
+      mode: mode,
+    ));
+
+    final demo = mode == LoginMode.demoCore;
+    final coreUrl = demo ? demoCoreUrlFromEnvironment : coreUrlFromEnvironment;
+    if (coreUrl != null) {
+      await _verifyCoreVersionAndRetrieveSupportedLoginTypesSubmitted(coreUrl, defaultTenantId, demo);
+    }
+  }
+
+  // LoginCoreUrlAssign
+
+  void coreUrlInputChanged(String value) {
     emit(state.copyWith(
       coreUrlInput: UrlInput.dirty(value),
     ));
   }
 
   void loginCoreUrlAssignSubmitted() async {
-    if (!state.status.isInput || !state.coreUrlInput.isValid) {
+    if (state.processing || !state.coreUrlInput.isValid) {
       return;
     }
-
-    emit(state.copyWith(
-      status: LoginStatus.processing,
-    ));
 
     var coreUrlInputValue = state.coreUrlInput.value;
     if (!coreUrlInputValue.startsWith(RegExp(r'(https|http)://'))) {
       coreUrlInputValue = 'https://$coreUrlInputValue';
     }
-    final tenantId = defaultTenantId;
-    try {
-      final client = createWebtritApiClient(coreUrlInputValue, tenantId);
-      await _verifyCoreVersion(client);
-      emit(state.copyWith(
-        status: LoginStatus.ok,
-        coreUrl: coreUrlInputValue,
-        tenantId: tenantId,
-      ));
-    } catch (e) {
-      emit(state.copyWith(
-        status: LoginStatus.input,
-        error: e,
-      ));
-    }
+
+    await _verifyCoreVersionAndRetrieveSupportedLoginTypesSubmitted(coreUrlInputValue, defaultTenantId);
   }
 
   void loginCoreUrlAssignBack() async {
-    if (state.status != LoginStatus.input) {
-      return;
-    }
-
     emit(state.copyWith(
-      status: LoginStatus.back,
+      mode: null,
       coreUrlInput: const UrlInput.pure(),
     ));
   }
 
-  // LoginOtpRequestTab
+  // LoginSwitch
 
-  void loginOptRequestEmailInputChanged(String value) {
+  void loginSwitchBack() async {
     emit(state.copyWith(
-      emailInput: EmailInput.dirty(value),
+      mode: state.mode == LoginMode.customCore ? state.mode : null,
+      coreUrl: null,
+      tenantId: null,
+      supportedLoginTypes: null,
+      otpSigninSessionOtpProvisionalWithDateTime: null,
+      passwordSigninPasswordInputObscureText: true,
+      signupSessionOtpProvisionalWithDateTime: null,
+      otpSigninUserRefInput: const UserRefInput.pure(),
+      otpSigninCodeInput: const CodeInput.pure(),
+      passwordSigninUserRefInput: const UserRefInput.pure(),
+      passwordSigninPasswordInput: const PasswordInput.pure(),
+      signupEmailInput: const EmailInput.pure(),
+      signupCodeInput: const CodeInput.pure(),
     ));
   }
 
-  void loginOptRequestPhoneInputChanged(String value) {
+  // LoginOtpSigninRequest
+
+  void otpSigninUserRefInputChanged(String value) {
     emit(state.copyWith(
-      phoneInput: PhoneInput.dirty(value),
+      otpSigninUserRefInput: UserRefInput.dirty(value),
     ));
   }
 
-  void loginOptRequestSubmitted() async {
-    if (!state.status.isInput || !(state.demo ? state.emailInput.isValid : state.phoneInput.isValid)) {
+  void loginOptSigninRequestSubmitted() async {
+    if (state.processing || !state.otpSigninUserRefInput.isValid) {
       return;
     }
 
     emit(state.copyWith(
-      status: LoginStatus.processing,
+      processing: true,
     ));
+
     try {
-      final SessionResult result;
       final client = createWebtritApiClient(state.coreUrl!, state.tenantId!);
-      if (state.demo) {
-        result = await _createUserRequest(client, state.emailInput.value);
-      } else {
-        result = await _sessionOtpRequest(client, state.phoneInput.value);
-      }
-      if (result is SessionOtpProvisional) {
-        emit(state.copyWith(
-          status: LoginStatus.ok,
-          sessionOtpProvisional: result,
-        ));
-      } else if (result is SessionToken) {
-        // TODO: Add logic for navigate authorized user
-        throw UnimplementedError();
-      } else {
-        // TODO: Add logic for navigate to undefine page
-        throw UnimplementedError();
-      }
+      final sessionOtpProvisional = await _createSessionOtp(client, state.otpSigninUserRefInput.value);
+
+      emit(state.copyWith(
+        processing: false,
+        otpSigninSessionOtpProvisionalWithDateTime: (sessionOtpProvisional, DateTime.now()),
+      ));
     } catch (e) {
       emit(state.copyWith(
-        status: LoginStatus.input,
+        processing: false,
         error: e,
       ));
     }
   }
 
-  void loginOptRequestBack() async {
-    if (state.status != LoginStatus.input) {
+  // LoginOtpSigninVerify
+
+  void otpSigninCodeInputChanged(String value) {
+    emit(state.copyWith(
+      otpSigninCodeInput: CodeInput.dirty(value),
+    ));
+  }
+
+  void loginOptSigninVerifySubmitted() async {
+    if (state.processing || !state.otpSigninCodeInput.isValid) {
       return;
     }
 
     emit(state.copyWith(
-      status: LoginStatus.back,
-      coreUrl: null,
-      tenantId: null,
-      emailInput: const EmailInput.pure(),
-      phoneInput: const PhoneInput.pure(),
-    ));
-  }
-
-  // LoginOtpVerifyTab
-
-  void loginOptVerifyCodeInputChanged(String value) {
-    emit(state.copyWith(
-      codeInput: CodeInput.dirty(value),
-    ));
-  }
-
-  void loginOptVerifySubmitted() async {
-    if (state.status != LoginStatus.input || !state.codeInput.isValid) {
-      return;
-    }
-
-    emit(state.copyWith(
-      status: LoginStatus.processing,
+      processing: true,
     ));
     try {
       final client = createWebtritApiClient(state.coreUrl!, state.tenantId!);
-      final sessionToken = await _sessionOtpVerify(client, state.sessionOtpProvisional!, state.codeInput.value);
+      final sessionToken = await _verifySessionOtp(
+          client, state.otpSigninSessionOtpProvisionalWithDateTime!.$1, state.otpSigninCodeInput.value);
 
+      // does not set processing to false to hold processing widgets state during navigation
       emit(state.copyWith(
-        status: LoginStatus.ok,
         tenantId: sessionToken.tenantId ?? state.tenantId!,
         token: sessionToken.token,
       ));
     } catch (e) {
       emit(state.copyWith(
-        status: LoginStatus.input,
+        processing: false,
         error: e,
       ));
     }
   }
 
-  void loginOptVerifyBack() async {
-    if (state.status != LoginStatus.input) {
-      return;
-    }
-
+  void loginOptSigninVerifyBack() async {
     emit(state.copyWith(
-      status: LoginStatus.back,
-      sessionOtpProvisional: null,
-      codeInput: const CodeInput.pure(),
+      otpSigninSessionOtpProvisionalWithDateTime: null,
+      otpSigninCodeInput: const CodeInput.pure(),
     ));
   }
 
-  void loginOptVerifyRepeat() async {
-    if (state.status != LoginStatus.input) {
+  void loginOptSigninVerifyRepeat() {
+    loginOptSigninRequestSubmitted();
+  }
+
+  // LoginPasswordSignin
+
+  void passwordSigninUserRefInputChanged(String value) {
+    emit(state.copyWith(
+      passwordSigninUserRefInput: UserRefInput.dirty(value),
+    ));
+  }
+
+  void passwordSigninPasswordInputChanged(String value) {
+    emit(state.copyWith(
+      passwordSigninPasswordInput: PasswordInput.dirty(value),
+    ));
+  }
+
+  void passwordSigninPasswordInputObscureTextToggled() {
+    emit(state.copyWith(
+      passwordSigninPasswordInputObscureText: !state.passwordSigninPasswordInputObscureText,
+    ));
+  }
+
+  void loginPasswordSigninSubmitted() async {
+    if (state.processing || !state.passwordSigninUserRefInput.isValid || !state.passwordSigninPasswordInput.isValid) {
       return;
     }
 
     emit(state.copyWith(
-      status: LoginStatus.processing,
+      processing: true,
     ));
 
-    // TODO: Part of duplicate code with method loginOptRequestSubmitted
     try {
-      final SessionResult result;
       final client = createWebtritApiClient(state.coreUrl!, state.tenantId!);
-      if (state.demo) {
-        result = await _createUserRequest(client, state.emailInput.value);
-      } else {
-        result = await _sessionOtpRequest(client, state.phoneInput.value);
-      }
+      final sessionToken = await _createSessionRequest(
+          client, state.passwordSigninUserRefInput.value, state.passwordSigninPasswordInput.value);
+
+      // does not set processing to false to hold processing widgets state during navigation
+      emit(state.copyWith(
+        tenantId: sessionToken.tenantId ?? state.tenantId!,
+        token: sessionToken.token,
+      ));
+    } catch (e) {
+      emit(state.copyWith(
+        processing: false,
+        error: e,
+      ));
+    }
+  }
+
+  void loginPasswordSigninBack() async {
+    emit(state.copyWith(
+      passwordSigninUserRefInput: const UserRefInput.pure(),
+      passwordSigninPasswordInput: const PasswordInput.pure(),
+    ));
+  }
+
+  // LoginSignupRequest
+
+  void signupEmailInputChanged(String value) {
+    emit(state.copyWith(
+      signupEmailInput: EmailInput.dirty(value),
+    ));
+  }
+
+  void loginSignupRequestSubmitted() async {
+    if (state.processing || !state.signupEmailInput.isValid) {
+      return;
+    }
+
+    emit(state.copyWith(
+      processing: true,
+    ));
+
+    try {
+      final client = createWebtritApiClient(state.coreUrl!, state.tenantId!);
+      final result = await _createUserRequest(client, state.signupEmailInput.value);
+
       if (result is SessionOtpProvisional) {
         emit(state.copyWith(
-          status: LoginStatus.input,
-          sessionOtpProvisional: result,
+          processing: false,
+          signupSessionOtpProvisionalWithDateTime: (result, DateTime.now()),
         ));
       } else if (result is SessionToken) {
-        // TODO: Add logic for navigate authorized user
-        throw UnimplementedError();
+        // does not set processing to false to hold processing widgets state during navigation
+        emit(state.copyWith(
+          tenantId: result.tenantId ?? state.tenantId!,
+          token: result.token,
+        ));
       } else {
-        // TODO: Add logic for navigate to undefine page
         throw UnimplementedError();
       }
     } catch (e) {
       emit(state.copyWith(
-        status: LoginStatus.input,
+        processing: false,
         error: e,
       ));
     }
   }
+
+  // LoginSignupVerify
+
+  void signupCodeInputChanged(String value) {
+    emit(state.copyWith(
+      signupCodeInput: CodeInput.dirty(value),
+    ));
+  }
+
+  void loginSignupVerifySubmitted() async {
+    if (state.processing || !state.signupCodeInput.isValid) {
+      return;
+    }
+
+    emit(state.copyWith(
+      processing: true,
+    ));
+
+    try {
+      final client = createWebtritApiClient(state.coreUrl!, state.tenantId!);
+      final sessionToken = await _verifySessionOtp(
+          client, state.signupSessionOtpProvisionalWithDateTime!.$1, state.signupCodeInput.value);
+
+      // does not set processing to false to hold processing widgets state during navigation
+      emit(state.copyWith(
+        tenantId: sessionToken.tenantId ?? state.tenantId!,
+        token: sessionToken.token,
+      ));
+    } catch (e) {
+      emit(state.copyWith(
+        processing: false,
+        error: e,
+      ));
+    }
+  }
+
+  void loginSignupVerifyBack() async {
+    emit(state.copyWith(
+      signupSessionOtpProvisionalWithDateTime: null,
+      signupCodeInput: const CodeInput.pure(),
+    ));
+  }
+
+  void loginSignupVerifyRepeat() {
+    loginSignupRequestSubmitted();
+  }
 }
 
-Future<void> _verifyCoreVersion(
+Future<List<LoginType>?> _verifyCoreVersionAndRetrieveSupportedLoginTypes(
   WebtritApiClient webtritApiClient,
 ) async {
-  final actualCoreVersion = (await webtritApiClient.getSystemInfo()).core.version;
-  CoreVersion.supported().verify(actualCoreVersion);
+  final systemInfo = await webtritApiClient.getSystemInfo();
+
+  CoreVersion.supported().verify(systemInfo.core.version);
+
+  final supportedLoginTypeNames = Set.from(LoginType.values.map((e) => e.name));
+  return systemInfo.adapter?.supported
+      ?.where((supportedName) => supportedLoginTypeNames.contains(supportedName))
+      .map((supportedLoginTypeName) => LoginType.values.byName(supportedLoginTypeName))
+      .toList();
+}
+
+Future<SessionOtpProvisional> _createSessionOtp(
+  WebtritApiClient webtritApiClient,
+  String userRef,
+) async {
+  return await webtritApiClient.createSessionOtp(SessionOtpCredential(
+    bundleId: PackageInfo().packageName,
+    type: PlatformInfo().appType,
+    identifier: AppInfo().identifier,
+    userRef: userRef,
+  ));
+}
+
+Future<SessionToken> _verifySessionOtp(
+  WebtritApiClient webtritApiClient,
+  SessionOtpProvisional sessionOtpProvisional,
+  String code,
+) async {
+  return await webtritApiClient.verifySessionOtp(sessionOtpProvisional, code);
+}
+
+Future<SessionToken> _createSessionRequest(
+  WebtritApiClient webtritApiClient,
+  String userRef,
+  String password,
+) async {
+  return await webtritApiClient.createSession(SessionLoginCredential(
+    bundleId: PackageInfo().packageName,
+    type: PlatformInfo().appType,
+    identifier: AppInfo().identifier,
+    login: userRef,
+    password: password,
+  ));
 }
 
 Future<SessionResult> _createUserRequest(
@@ -325,22 +429,6 @@ Future<SessionResult> _createUserRequest(
   ));
 }
 
-Future<SessionOtpProvisional> _sessionOtpRequest(
-  WebtritApiClient webtritApiClient,
-  String phone,
-) async {
-  return await webtritApiClient.createSessionOtp(SessionOtpCredential(
-    bundleId: PackageInfo().packageName,
-    type: PlatformInfo().appType,
-    identifier: AppInfo().identifier,
-    userRef: phone,
-  ));
-}
+abstract class LoginCubitException implements Exception {}
 
-Future<SessionToken> _sessionOtpVerify(
-  WebtritApiClient webtritApiClient,
-  SessionOtpProvisional sessionOtpProvisional,
-  String code,
-) async {
-  return await webtritApiClient.verifySessionOtp(sessionOtpProvisional, code);
-}
+class SupportedLoginTypeMissedException extends LoginCubitException {}
