@@ -1,40 +1,34 @@
 import 'package:flutter/material.dart';
 
-class CountDownBuilder extends StatefulWidget {
+class CountDownBuilder extends StatelessWidget {
   const CountDownBuilder({
     super.key,
+    required this.start,
     required this.interval,
     required this.builder,
   });
 
+  final DateTime start;
   final Duration interval;
   final Widget Function(BuildContext context, int counter) builder;
 
   @override
-  State<CountDownBuilder> createState() => _CountDownBuilderState();
-}
-
-class _CountDownBuilderState extends State<CountDownBuilder> {
-  late Stream<int> intervalStream;
-
-  @override
-  void initState() {
-    super.initState();
-
-    intervalStream = Stream.periodic(const Duration(seconds: 1), (i) => widget.interval.inSeconds - i - 1)
-        .take(widget.interval.inSeconds);
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return StreamBuilder<int>(
-      stream: intervalStream,
-      builder: (context, snapshot) {
-        final seconds = snapshot.connectionState == ConnectionState.waiting
-            ? widget.interval.inSeconds
-            : snapshot.data ?? widget.interval.inSeconds;
-        return widget.builder(context, seconds);
-      },
-    );
+    final elapsed = DateTime.now().difference(start);
+
+    var remain = interval - elapsed;
+    if (remain <= Duration.zero) {
+      return builder(context, 0);
+    } else {
+      return StreamBuilder<int>(
+        stream: Stream.periodic(const Duration(seconds: 1), (i) => remain.inSeconds - i - 1).take(remain.inSeconds),
+        builder: (context, snapshot) {
+          final seconds = snapshot.connectionState == ConnectionState.waiting
+              ? remain.inSeconds
+              : snapshot.data ?? remain.inSeconds;
+          return builder(context, seconds);
+        },
+      );
+    }
   }
 }
