@@ -73,6 +73,9 @@ class AppLocalizationsUk extends AppLocalizations {
   String get call_description_outgoing => 'Вихідний дзвінок';
 
   @override
+  String get call_description_transferProcessing => 'Обробка переадресації';
+
+  @override
   String get call_FailureAcknowledgeDialog_title => 'Помилка';
 
   @override
@@ -130,6 +133,9 @@ class AppLocalizationsUk extends AppLocalizations {
   String get contactsSourceLocal => 'Ваш телефон';
 
   @override
+  String get contacts_Text_blingTransferInitiated => 'Сліпе переадресування';
+
+  @override
   String get copyToClipboard_floatingSnackBar => 'Текст скопійовано';
 
   @override
@@ -171,6 +177,9 @@ class AppLocalizationsUk extends AppLocalizations {
   }
 
   @override
+  String get favorites_Text_blingTransferInitiated => 'Сліпе переадресування';
+
+  @override
   String get locale_default => 'За замовчуванням';
 
   @override
@@ -180,39 +189,47 @@ class AppLocalizationsUk extends AppLocalizations {
   String get locale_uk => 'Українська';
 
   @override
-  String get login_AppBarTitle_coreUrlAssign => '';
-
-  @override
-  String get login_AppBarTitle_otpRequest => '';
-
-  @override
-  String get login_AppBarTitle_otpVerify => '';
-
-  @override
   String get login_Button_coreUrlAssignProceed => 'Продовжити';
 
   @override
-  String get login_Button_otpRequestProceed => 'Продовжити';
+  String get login_Button_otpSigninRequestProceed => 'Продовжити';
 
   @override
-  String get login_Button_otpVerifyProceed => 'Перевірити';
+  String get login_Button_otpSigninVerifyProceed => 'Перевірити';
 
   @override
-  String get login_Button_otpVerifyRepeat => 'Повторно надіслати код';
+  String get login_Button_otpSigninVerifyRepeat => 'Повторно надіслати код';
 
   @override
-  String login_Button_otpVerifyRepeatInterval(int seconds) {
+  String login_Button_otpSigninVerifyRepeatInterval(int seconds) {
     return 'Повторно надіслати код ($seconds с)';
   }
+
+  @override
+  String get login_Button_passwordSigninProceed => 'Продовжити';
 
   @override
   String get login_Button_signIn => 'Увійти';
 
   @override
+  String get login_Button_signupRequestProceed => 'Продовжити';
+
+  @override
   String get login_Button_signUpToDemoInstance => 'Зареєструватися';
 
   @override
-  String get login_ButtonTooltip_signInToYourInstance => 'Увійдіть до свогоWebTrit Cloud Backend';
+  String get login_Button_signupVerifyProceed => 'Перевірити';
+
+  @override
+  String get login_Button_signupVerifyRepeat => 'Повторно надіслати код';
+
+  @override
+  String login_Button_signupVerifyRepeatInterval(int seconds) {
+    return 'Повторно надіслати код ($seconds с)';
+  }
+
+  @override
+  String get login_ButtonTooltip_signInToYourInstance => 'Увійдіть до свого WebTrit Cloud Backend';
 
   @override
   String login_CoreVersionUnsupportedExceptionError(String actual, String supportedConstraint) {
@@ -238,10 +255,16 @@ class AppLocalizationsUk extends AppLocalizations {
   String get login_RequestFailureOtpVerificationAttemptsExceededError => 'Перевищено кількість спроб підтвердження';
 
   @override
+  String get login_RequestFailureParametersApplyIssueError => 'Надані дані не можуть бути оброблені';
+
+  @override
   String get login_RequestFailurePhoneNotFoundError => 'Номер телефону не знайдено';
 
   @override
   String get login_RequestFailureUnconfiguredBundleIdError => 'Помилка конфігурації сервера додатка - сповістіть свого постачальника послуг';
+
+  @override
+  String get login_SupportedLoginTypeMissedExceptionError => 'Поточний WebTrit Cloud Backend не підтримує жодного типу входу, сумісного з цим додатком';
 
   @override
   String login_Text_coreUrlAssignPostDescription(Object email) {
@@ -255,37 +278,86 @@ class AppLocalizationsUk extends AppLocalizations {
   String get login_TextFieldLabelText_coreUrlAssign => 'Введіть URL вашого WebTrit Cloud Backend';
 
   @override
-  String get login_TextFieldLabelText_otpRequestEmail => 'Введіть свою електронну пошту';
+  String get login_TextFieldLabelText_otpSigninCode => 'Введіть код підтвердження';
 
   @override
-  String get login_TextFieldLabelText_otpRequestPhone => 'Введіть свій номер телефону';
+  String get login_TextFieldLabelText_otpSigninUserRef => 'Введіть свій номер телефону чи електронну пошту';
 
   @override
-  String get login_TextFieldLabelText_otpVerifyCode => 'Введіть код підтвердження';
+  String get login_TextFieldLabelText_passwordSigninPassword => 'Введіть ваш пароль';
 
   @override
-  String get login_Text_otpRequestDemoDescription => 'Якщо у вас ще немає облікового запису, він буде автоматично створений для вас.';
+  String get login_TextFieldLabelText_passwordSigninUserRef => 'Введіть свій номер телефону чи електронну пошту';
 
   @override
-  String get login_Text_otpRequestDescription => '';
+  String get login_TextFieldLabelText_signupCode => 'Введіть код підтвердження';
 
   @override
-  String login_Text_otpVerifyCheckSpamFrom(String email) {
+  String get login_TextFieldLabelText_signupEmail => 'Введіть свою електронну пошту';
+
+  @override
+  String get login_Text_otpSigninRequestPostDescription => '';
+
+  @override
+  String get login_Text_otpSigninRequestPreDescription => '';
+
+  @override
+  String login_Text_otpSigninVerifyPostDescriptionFromEmail(String email) {
     return 'Якщо ви не бачите електронного листа з кодом підтвердження від $email у своїй скринці вхідних, будь ласка, перевірте папку спаму.';
   }
 
   @override
-  String get login_Text_otpVerifyCheckSpamGeneral => 'Якщо ви не бачите електронного листа з кодом підтвердження у своїй скринці вхідних, будь ласка, перевірте папку спаму.';
+  String get login_Text_otpSigninVerifyPostDescriptionGeneral => 'Якщо ви не бачите електронного листа з кодом підтвердження у своїй скринці вхідних, будь ласка, перевірте папку спаму.';
 
   @override
-  String login_Text_otpVerifySentToEmail(String email) {
-    return 'На $email надіслано одноразовий код підтвердження.';
+  String login_Text_otpSigninVerifyPreDescriptionUserRef(String userRef) {
+    return 'На електронну пошту, прив\'язану до наданого номера телефону чи електронної пошти, надіслано одноразовий код підтвердження.';
   }
 
   @override
   String login_Text_otpVerifySentToEmailAssignedWithPhone(String phone) {
     return 'На електронну пошту, прив\'язану до телефону tel:$phone, надіслано одноразовий код підтвердження.';
   }
+
+  @override
+  String get login_Text_passwordSigninPostDescription => '';
+
+  @override
+  String get login_Text_passwordSigninPreDescription => '';
+
+  @override
+  String get login_Text_signupRequestPostDescription => '';
+
+  @override
+  String get login_Text_signupRequestPostDescriptionDemo => 'Якщо у вас ще немає облікового запису, він буде автоматично створений для вас.';
+
+  @override
+  String get login_Text_signupRequestPreDescription => '';
+
+  @override
+  String get login_Text_signupRequestPreDescriptionDemo => '';
+
+  @override
+  String login_Text_signupVerifyPostDescriptionFromEmail(String email) {
+    return 'Якщо ви не бачите електронного листа з кодом підтвердження від $email у своїй скринці вхідних, будь ласка, перевірте папку спаму.';
+  }
+
+  @override
+  String get login_Text_signupVerifyPostDescriptionGeneral => 'Якщо ви не бачите електронного листа з кодом підтвердження у своїй скринці вхідних, будь ласка, перевірте папку спаму.';
+
+  @override
+  String login_Text_signupVerifyPreDescriptionEmail(String email) {
+    return 'На $email надіслано одноразовий код підтвердження.';
+  }
+
+  @override
+  String get loginType_otpSignin => 'Увійти за OTP';
+
+  @override
+  String get loginType_passwordSignin => 'Вхід за паролем';
+
+  @override
+  String get loginType_signup => 'Зареєструватися';
 
   @override
   String get login_validationCoreUrlError => 'Будь ласка, введіть правильний URL';
@@ -295,6 +367,9 @@ class AppLocalizationsUk extends AppLocalizations {
 
   @override
   String get login_validationPhoneError => 'Будь ласка, введіть правильний номер телефону';
+
+  @override
+  String get login_validationUserRefError => 'Будь ласка, введіть правильний номер телефону чи електронну пошту';
 
   @override
   String get logRecordsConsole_AppBarTitle => 'Консоль журналу';
@@ -379,6 +454,9 @@ class AppLocalizationsUk extends AppLocalizations {
   String recents_snackBar_deleted(String name) {
     return '$name видалено';
   }
+
+  @override
+  String get recents_Text_blingTransferInitiated => 'Сліпе переадресування';
 
   @override
   String get recentsVisibilityFilter_all => 'Всі';
