@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'package:auto_route/auto_route.dart';
+
 class ExtBackButton extends StatelessWidget {
   const ExtBackButton({
     super.key,
@@ -10,12 +12,31 @@ class ExtBackButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return IconButton(
-      icon: const Icon(
-        Icons.arrow_back,
-      ),
-      tooltip: MaterialLocalizations.of(context).backButtonTooltip,
-      onPressed: disabled ? null : () => Navigator.maybePop(context),
+    return AutoLeadingButton(
+      builder: (context, leadingType, action) {
+        switch (leadingType) {
+          case LeadingType.back:
+            return IconButton(
+              icon: const BackButtonIcon(),
+              tooltip: MaterialLocalizations.of(context).backButtonTooltip,
+              onPressed: disabled ? null : action,
+            );
+          case LeadingType.close:
+            return IconButton(
+              icon: const BackButtonIcon(),
+              tooltip: MaterialLocalizations.of(context).closeButtonTooltip,
+              onPressed: disabled ? null : action,
+            );
+          case LeadingType.drawer:
+            return IconButton(
+              icon: const BackButtonIcon(),
+              tooltip: MaterialLocalizations.of(context).openAppDrawerTooltip,
+              onPressed: disabled ? null : action,
+            );
+          case LeadingType.noLeading:
+            return const SizedBox.shrink();
+        }
+      },
     );
   }
 }
