@@ -79,7 +79,7 @@ void main() async {
     takeScreenshotTestWidgets('login_screen__modeSelect', () {
       return ScreenshotApp(
         appBloc: appBloc,
-        child: const LoginScreenScreenshot(LoginStep.modeSelect),
+        child: const LoginModeSelectScreenScreenshot(),
       );
     });
   });
