@@ -1,4 +1,4 @@
 export 'call_screen_screenshot.dart';
-export 'login_screen_screenshot.dart';
+export 'login_mode_select_screen_screenshot.dart';
 export 'main_screen_screenshot.dart';
 export 'setting_screen_screenshot.dart';
